fix: Update list of collected artists when artist gets added or removed
